Update sketch.js
i started to tweak some code that I already had. WE are figuring out the details, but the base is coming together well. If I wanted to make the ball disapear is i catch it, how would I do that? Also do I have to drak the book shelf in the background or is there another way?